ユーザ登録ページとログインページのテストを個別のファイルに分割
